+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.1] - 2020-02-20
+
 ### Fixed
 
 - Extract Variable didn't worked properly with empty string and special keywords. This is now fixed!
@@ -711,7 +713,8 @@ function doSomething(someData) {
 
 </details>
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/3.2.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/3.2.1...HEAD
+[3.2.1]: https://github.com/nicoespeon/abracadabra/compare/3.2.0...3.2.1
 [3.2.0]: https://github.com/nicoespeon/abracadabra/compare/3.1.0...3.2.0
 [3.1.0]: https://github.com/nicoespeon/abracadabra/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/nicoespeon/abracadabra/compare/2.0.0...3.0.0